Debugging role-based classification

+++ src/misc/matrix.py
@@ -6,7 +6,7 @@ def apply_permutation(module, X, permutation):
 def pair_range(x, y, dx=1, dy=1):
     ax = torch.arange(x).view(1, x).repeat(y, 1).view(-1) * dx
     ay = torch.arange(y).view(y, 1).repeat(1, x).view(-1) * dy
-    return ax, ay
+    return ax.long(), ay.long()
 
 def true_permute(v1, v2):
     w1 = v1.view(-1, 1).repeat(1, len(v2)).view(-1)
